+++ JavaScriptRuntime/Boolean.cs
@@ -1,0 +1,7 @@
+namespace JavaScriptRuntime
+{
+    [IntrinsicObject("Boolean")]
+    public static class Boolean
+    {
+    }
+}